BasePersistenceTest is removed
