+++ legacy-data-access/src/test/resources/META-INF/persistence.xml
@@ -1,0 +1,104 @@
+<persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence
+             http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd"  version="2.1">
+
+  <persistence-unit name="CWSCMS-TEST" transaction-type="RESOURCE_LOCAL">
+    <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+
+    <class>gov.ca.cwds.data.legacy.cms.entity.BackgroundCheck</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.Case</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CaseAssignment</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CaseLoad</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CaseLoadWeighting</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ChildClient</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.Client</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ClientPaternityDetail</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ClientRelationship</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.OtherEthnicity</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ClientOtherEthnicity</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ScpOtherEthnicity</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CollateralInividualPaternityDetail</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CountyLicenseCase</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CreditReportHistory</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.CsecHistory</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.DasHistory</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.DeliveredService</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.FCEligibility</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.HealthInterventionPlan</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.LicensingVisit</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.LongText</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.MedicalEligibilityApplication</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.NearFatality</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.OtherAdultsInPlacementHome</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.OtherChildrenInPlacementHome</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.OtherPeopleScpRelationship</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.OutOfHomePlacement</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.OutOfStateCheck</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ParentalRightsTermination</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.PaternityDetail</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.PlacementEpisode</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.PlacementHome</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.PlacementHomeNotes</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.PlacementHomeProfile</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.Referral</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ReferralAssignment</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.ReferralClient</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.SafetyAlert</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.SpecialEducation</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.StaffPerson</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.StaffPersonCaseLoad</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.SubstituteCareProvider</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.HealthReferral</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.TribalAncestryNotification</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.HealthScreening</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.SchoolOriginHistory</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.PlacementFacilityTypeHistory</class>
+
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.ActiveServiceComponentType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.ApprovalStatusType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.CaseClosureReasonType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.ClientRelationshipType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.Country</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.County</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.CreditAgencyType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.CreditReportRequestedByType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.CreditReportStatusType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.DeathCircumstancesType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.Ethnicity</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.FacilityType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.FcEligibilityDentalReason</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.FcEligibilityTermReason</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthPlanTerminationType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.ImmigrationStatus</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.Language</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.LicenseStatus</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.MaritalStatus</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.NameType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.ParentalRightTerminationType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.Religion</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.SafetyAlertActivationReasonType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.SecondaryAssignmentRoleType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.ServiceContactType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.SexualExploitationType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.State</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.SystemCode</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.VisitType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthConsentType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthReferredToType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthReferralType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthScreeningType</class>
+    <class>gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthScreenedByType</class>
+
+    <properties>
+      <property name="hibernate.current_session_context_class" value="thread"/>
+      <property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect"/>
+      <property name="hibernate.connection.driver_class" value="org.h2.Driver"/>
+      <property name="hibernate.show_sql" value="true"/>
+      <property name="hibernate.connection.url" value="jdbc:h2:mem:cwscms;INIT=create schema if not exists CWSCMS\;set schema CWSCMS;DB_CLOSE_DELAY=-1"/>
+      <property name="hibernate.connection.username" value=""/>
+      <property name="hibernate.connection.password" value=""/>
+      <property name="hibernate.default_schema" value="CWSCMS"/>
+    </properties>
+  </persistence-unit>
+</persistence>
