--- legacy-data-access/src/test/resources/hibernate.cfg.xml
@@ -1,96 +1,0 @@
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-  <session-factory>
-    <property name="hibernate.current_session_context_class">thread</property>
-    <property name="hibernate.dialect">org.hibernate.dialect.H2Dialect</property>
-    <property name="hibernate.connection.driver_class">org.h2.Driver</property>
-    <property name="show_sql">true</property>
-
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.BackgroundCheck" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.Case" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CaseAssignment" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CaseLoad" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CaseLoadWeighting" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ChildClient" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.Client" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ClientPaternityDetail"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ClientRelationship"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.OtherEthnicity" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ClientOtherEthnicity" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ScpOtherEthnicity" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CollateralInividualPaternityDetail"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CountyLicenseCase" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CreditReportHistory"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.CsecHistory"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.DasHistory"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.DeliveredService"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.FCEligibility" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.HealthInterventionPlan"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.LicensingVisit" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.LongText" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.MedicalEligibilityApplication" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.NearFatality"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.OtherAdultsInPlacementHome" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.OtherChildrenInPlacementHome" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.OtherPeopleScpRelationship" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.OutOfHomePlacement" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.OutOfStateCheck" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ParentalRightsTermination"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.PaternityDetail"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.PlacementEpisode" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.PlacementHome" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.PlacementHomeNotes" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.PlacementHomeProfile" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.Referral" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ReferralAssignment" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.ReferralClient" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.SafetyAlert" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.SpecialEducation" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.StaffPerson" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.StaffPersonCaseLoad" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.SubstituteCareProvider" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.HealthReferral"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.TribalAncestryNotification"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.HealthScreening"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.SchoolOriginHistory"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.PlacementFacilityTypeHistory"/>
-
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ActiveServiceComponentType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ApprovalStatusType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.CaseClosureReasonType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ClientRelationshipType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.Country"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.County"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.CreditAgencyType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.CreditReportRequestedByType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.CreditReportStatusType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.DeathCircumstancesType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.Ethnicity"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.FacilityType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.FcEligibilityDentalReason"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.FcEligibilityTermReason"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthPlanTerminationType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ImmigrationStatus"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.Language"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.LicenseStatus"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.MaritalStatus"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.NameType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ParentalRightTerminationType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.Religion"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.SafetyAlertActivationReasonType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.SecondaryAssignmentRoleType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ServiceContactType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.SexualExploitationType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.State"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.SystemCode"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.VisitType"/>
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthConsentType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthReferredToType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthReferralType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthScreeningType" />
-    <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.HealthScreenedByType" />
-  </session-factory>
-</hibernate-configuration>
